доработки по замечаниям
