+++ pom.xml
@@ -19,9 +19,16 @@
         <aspectj.version>1.9.22.1</aspectj.version>
         <allure-maven.version>2.14.0</allure-maven.version>
         <report.version>2.30.0</report.version>
+        <lombok.version>1.18.34</lombok.version>
     </properties>
 
     <dependencies>
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <version>${lombok.version}</version>
+            <scope>provided</scope>
+        </dependency>
         <dependency>
             <groupId>io.rest-assured</groupId>
             <artifactId>rest-assured</artifactId>
