+++ src/test/java/CreateOrderTests.java
@@ -1,5 +1,5 @@
 import apisteps.OrderApi;
-import POJO.UserData;
+import pojo.UserData;
 import apisteps.UserApi;
 import io.qameta.allure.Description;
 import io.qameta.allure.restassured.AllureRestAssured;
@@ -14,7 +14,7 @@ import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.notNullValue;
 
 public class CreateOrderTests {
-    private final String BASE_URL = "https://stellarburgers.nomoreparties.site";
+
     private UserApi userApi;
     public String userEmail = "dddddeddaas@yandex.ru";
     public String userName = "Magomed";
@@ -23,7 +23,7 @@ public class CreateOrderTests {
 
     @Before
     public void setUp() {
-        RestAssured.baseURI = BASE_URL;
+        RestAssured.baseURI = UserApi.baseUrl;
         RestAssured.filters(new AllureRestAssured());
         userApi = new UserApi();
         UserData userData = new UserData(userEmail, userPassword, userName);
