+++ src/test/java/GetUserOrdersTests.java
@@ -1,4 +1,4 @@
-import POJO.UserData;
+import pojo.UserData;
 import apisteps.OrderApi;
 import apisteps.UserApi;
 import io.qameta.allure.Description;
@@ -14,7 +14,7 @@ import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.notNullValue;
 
 public class GetUserOrdersTests {
-    private final String BASE_URL = "https://stellarburgers.nomoreparties.site";
+
     private UserApi userApi;
     public String userEmail = "dddddeddaas@yandex.ru";
     public String userName = "Magomed";
@@ -24,7 +24,7 @@ public class GetUserOrdersTests {
     @Before
     public void setUp() {
 // регистрация пользователя
-        RestAssured.baseURI = BASE_URL;
+        RestAssured.baseURI = UserApi.baseUrl;
         RestAssured.filters(new AllureRestAssured());
         userApi = new UserApi();
         UserData userData = new UserData(userEmail, userPassword, userName);
@@ -36,7 +36,6 @@ public class GetUserOrdersTests {
 // создание заказа пользователя
         OrderApi orderData = orderIngredients();
         userApi.createOrderDataWithToken(orderData, userAccessToken);
-
     }
 
     @After
