--- src/test/java/POJO/UserData.java
@@ -1,49 +1,0 @@
-package POJO;
-
-public class UserData {
-    private String email;
-    private String password;
-    private String name;
-    private boolean success;
-
-    public UserData(String email, String password, String name) {
-        this.email = email;
-        this.password = password;
-        this.name = name;
-    }
-
-    public UserData() {
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public boolean getSuccess() {
-        return success;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public void setSuccess() {
-        this.success = success;
-    }
-}
